+++ src/index.ts
@@ -460,13 +460,8 @@ export function createForgoInstance(customEnv: any) {
       }
     }
 
-    syncStateAndProps(
-      forgoNode,
-      node,
-      true,
-      pendingAttachStates,
-      oldComponentState
-    );
+    syncAttrsAndState(forgoNode, node, true, pendingAttachStates);
+    remountComponents(pendingAttachStates, oldComponentState);
     return { nodes: [node] };
   }
 
@@ -624,16 +619,17 @@ export function createForgoInstance(customEnv: any) {
 
       const oldComponentState = getForgoState(targetElement)?.components;
 
-      renderChildNodes(targetElement);
-      unloadMarkedNodes(targetElement, pendingAttachStates);
-
-      syncStateAndProps(
+      syncAttrsAndState(
         forgoElement,
         targetElement,
         false,
-        pendingAttachStates,
-        oldComponentState
+        pendingAttachStates
       );
+
+      renderChildNodes(targetElement);
+      unloadMarkedNodes(targetElement, pendingAttachStates);
+
+      remountComponents(pendingAttachStates, oldComponentState);
 
       return { nodes: [targetElement] };
     }
@@ -652,15 +648,11 @@ export function createForgoInstance(customEnv: any) {
         forgoElement.props.ref.value = newElement;
       }
 
+      syncAttrsAndState(forgoElement, newElement, true, pendingAttachStates);
+
       renderChildNodes(newElement);
 
-      syncStateAndProps(
-        forgoElement,
-        newElement,
-        true,
-        pendingAttachStates,
-        undefined
-      );
+      remountComponents(pendingAttachStates, undefined);
 
       return { nodes: [newElement] };
     }
@@ -1041,18 +1033,14 @@ export function createForgoInstance(customEnv: any) {
     );
   }
 
-  /*
-    Sync component states and props between a newNode and an oldNode.
-  */
-  function syncStateAndProps(
-    forgoNode: ForgoNode,
-    node: ChildNode,
-    isNewNode: boolean,
+  /**
+   * Attach component states to DOM nodes and call unmount/mount lifecycle
+   * methods
+   */
+  function remountComponents(
     pendingAttachStates: NodeAttachedComponentState<any>[],
     oldComponentStates: NodeAttachedComponentState<any>[] | undefined
   ) {
-    attachProps(forgoNode, node, isNewNode, pendingAttachStates);
-
     if (oldComponentStates) {
       const indexOfFirstIncompatibleState = findIndexOfFirstIncompatibleState(
         pendingAttachStates,
@@ -1397,7 +1385,7 @@ export function createForgoInstance(customEnv: any) {
    * Attach props from the forgoElement onto the DOM node. We also need to attach
    * states from pendingAttachStates
    */
-  function attachProps(
+  function syncAttrsAndState(
     forgoNode: ForgoNode,
     node: ChildNode,
     isNewNode: boolean,